requirements.txt
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-ads==24.1.0
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
isodate==0.6.1
lxml==5.2.2
numpy==2.0.0
oauthlib==3.2.2
pandas==2.2.2
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
text-unidecode==1.3
tzdata==2024.1
urllib3==2.2.2
xmltodict==0.13.0
zeep==4.2.1
